Extract .eslintrc from brigade repo
We want to share our ESLint configuration between all of our projects to
help keep things consistent and make it easier to maintain our projects.

ESLint now allows you to publish an NPM package for configuration that
can be used via ESLint's extends option.

This commit simply extracts the .eslintrc that we have been using in the
brigade repo, removes the repo-specific configuration (globals), and
breaks the configuration up into two files--one for projects that don't
use React and another for projects that do use React.