Enable all recommended rules
When I upgraded this configuration to be compatible with ESLint 1.0.0,
which disables all rules by default, I checked all of the rules listed
in the migration guide to make sure we were up to snuff. However, this
list did not include the recommended rules, so we were running with
those disabled for a while. This commit enables all of these rules, as
currently listed as recommended on the master list of rules:

  http://eslint.org/docs/rules/