Turn off `no-return-assign` check
With eslint 2.0, this rule started to apply to arrow functions as well,
so suddenly this type of construction would yield an error

beforeEach(() => this.foo = bar);

The main rationale for the `no-return-assign` rule is to prevent people
from accidentally assigning something when they meant to return a
boolean, the argument being that `=` and `==` are easily mixed up [1].
Since we pretty much never use double-equals for comparison (instead we
use triple-equals to avoid type conversion [2]), it feels a little odd
to have to either introduce a block for the arrow body, or wrap in
parenthesis.

beforeEach(() => (this.foo = bar)); // this would be fine

beforeEach(() => {
  this.foo = bar; // this would be fine too
});

So I'm turning off the check.

[1]: http://eslint.org/docs/rules/no-return-assign
[2]: http://stackoverflow.com/questions/359494/does-it-matter-which-equals-operator-vs-i-use-in-javascript-comparisons