Add warning for misaligned JSX closing tag
Kevin just called me out for this style infraction. I'd like our tooling
to call me out in the future, since Kevin is far too busy to always be
watching. :eyes:

But on a serious note, I believe it is the role of tools like eslint to
remove as much subjectivity from our code review process as possible. I
was pleasantly surprised to find this linting option as a viable way to
enforce the rule that Kevin suggested.

Examples of how to do it:

```jsx
<Hello shortProp="bar" />

<Hello
  veryLongPropName="bar"
/>

<Hello
  foo="bar"
  >
  <p>...</p>
</Hello>
```