Standardize on no spaces inside arrays
We have 2266 instances of arrays without spaces, to only 19 instances
of arrays with spaces.

$ ag --js '\[[^\n\]]+\]' | wc -l
2266

$ ag --js '\[ [^\n\]]+\]' | wc -l
19

(the first is a pretty bad regex, lots of false positives, but is still
way more than 19)

In the spirit of there being no ambiguity about the right way to do
things and reducing style comments in code review, this enables
"array-bracket-spacing: never" at the warning level.

[Airbnb also does this.][1]

[1]: https://github.com/airbnb/javascript#whitespace--in-brackets