This adds `peerDependencies` to the package
For `eslint`, this makes sense since this package cannot be used
outside the context of an ESLint environment.

For the two plugins, their inclusion is to support users using
npm2, which unlike npm3 and above can fail to resolve certain
depencies specified here if they are also specified in other parts
of the parent package. These will be removed from `peerDependencies`
in version 4 of this package.