Adds an `engines` and `engineStrict` config
This adds an `engines` and `engineStrict` field to our config.
This means we officially support:

Node >= 5.0.0 (the first version that shipped with npm3)
npm >= 3.3.6 (the version that shipped with Node 5)

`enginesStrict` was set because we know that the previous commit,
removing most of the `peerDependencies` will break this package
when using npm2 or below.