Forbid `class` props in react
In the past, we've had a custom overcommit hook protecting against using
`class` instead of `className` as a prop. For React newcomers, this is a
common mistake to make, so it's good to have some protection in place.
Instead of the custom overcommit hook, we can instead just prevent this
through the react/forbid-component-props rule [1].

[1] https://github.com/yannickcr/eslint-plugin-react/blob/master/docs/rules/forbid-component-props.md